MongoDB - complete

+++ src/main/java/sk/upjs/ics/bakalarka/dao/ReportDao.java
@@ -17,7 +17,7 @@ public interface ReportDao {
 
     public List<GlucoseRange> getRangesByPatient(String patientName);
 
-    public List<Report> getPatientByDaytimeAndRangeHigh(String daytime, BigDecimal rangeHigh);
+    public List<Report> getPatientByDaytimeAndRangeHigh(String daytime, double rangeHigh);
 
     //public List<RangeHighPatternTypePatientInfo> getRangeHighPatternPatientBy(int rangeNoOfDays, BigDecimal rangeHigh);
 }
